integrating sampling choice

+++ src/Interfaces/TracingServiceInterface.php
@@ -12,11 +12,11 @@ interface TracingServiceInterface
      * @param string $spanName
      * @param int $spanKind
      * @param \OpenTracing\SpanContext|null $parent
-     * @param float|null $timestamp
+     * @param int|null $timestamp
      *
      * @return \Romanpravda\Laravel\Tracing\Interfaces\SpanInterface
      */
-    public function startSpan(string $spanName, int $spanKind, ?SpanContextInterface $parent = null, ?float $timestamp = null): SpanInterface;
+    public function startSpan(string $spanName, int $spanKind, ?SpanContextInterface $parent = null, ?int $timestamp = null): SpanInterface;
 
     /**
      * Check for current span existence.
@@ -35,11 +35,11 @@ interface TracingServiceInterface
     /**
      * Ending current span.
      *
-     * @param float|null $timestamp
+     * @param int|null $timestamp
      *
      * @return void
      */
-    public function endCurrentSpan(?float $timestamp = null): void;
+    public function endCurrentSpan(?int $timestamp = null): void;
 
     /**
      * Writing context to given data.
